Fix WithoutKey filter and rename rest of With* functions.
Also renamed for the sink:
With -> WithKey
Without -> WithoutKey
WithValues -> WithValue
WithoutValues -> WithoutValue

It looks more consistent. Though it is not obvious with params now.
WithKey accepts arbitrary number of keys. WithValue accepts arbitrary
number of values and it is ok too. But WithValue want the key as first
parameter. It need more usability research. So these functions may be
renamed again.